Sassy Parallax Banner
Define the following variables at the top of your main.scss file—:

$parallax-height: 100vh;
$center-margin: 0px auto;

Reference those variables inside the instances that use them.
Start to create the logic for a parallax effect. Inside of both id selectors for our background parallax, #landscape and #monkeys add the following properties:
  width: 100%;
  height: 100vh;
  perspective: 1px;
  transform: translateZ(-1px);
  overflow-x: hidden;
  overflow-y: auto;

Since you are including travel photography in your parallax, you make sure that the background image covers the entire div. You can do this by setting the webkit-background-size to a value cover. This property is vendor specific, so it means you will need to make a mixin with the following properties:
    background-size: cover;
    -webkit-background-size: cover;
    -moz-background-size: cover;
    -o-background-size: cover;

The mixin should takes one parameter, the size, and assign that to all of the vendor properties.

Task 4: At the top of your main.scss, make a new mixin with the rest of the parallax background properties.
@mixin parallax-background($file) {
  background: url($file) no-repeat center center fixed;
  @include background-size(cover);
}

Include the mixin inside the selectors of the both #landscape and #monkey, passing in a value of “https://content.codecademy.com/projects/sass/landscape.jpg" and “https://content.codecademy.com/projects/sass/monkeys.jpg" respectively.

Assign the h1 property a
transform: translateZ(-2px);
This is to ensure that the Z-axis may be at a different translate setting than the image in the foreground

Scroll around in the webpage! Observe the parallax effect. Tweak the transform effects as you please and play around with the values to create your own effects.